Compartmentalize test helper support
