+++ test/support/test_case.rb
@@ -1,0 +1,5 @@
+ActionController::TestCase.class_eval do
+  def setup
+    @routes = TestHelper::Routes
+  end
+end
